requirements.txt
python-mpv==1.0.5
toml==0.10.2
ytmusicapi==1.6.0
